Split RulesParser to two files
